app(): Get the Bottle object (WSGI callable) you are currently looking for.
The improved app() function is a factory for Bottle objects. Without a
parameter, it returns a default Bottle instance. Most of the decorators and
the run() function use app() this way. You can change the default by providing
a name parameter or a Bottle object. Any further calls to app() in the same
thread will use the new default. When called from within a running Bottle
application, app() defaults to the current application.